ai-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.2
torch>=2.0.0
pydantic>=1.10.0
python-dotenv>=0.19.0
requests>=2.28.0
# openai>=1.0.0  # Optional: only used if OPENAI_API_KEY is set
